LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/interfaces/HTMLElement/HTMLTrackElement/kind-expected.txt


PASS HTMLTrackElement.kind missing value 
FAIL HTMLTrackElement.kind invalid value in content attribute assert_equals: expected "metadata" but got "subtitles"
PASS HTMLTrackElement.kind content attribute uppercase 
FAIL HTMLTrackElement.kind content attribute with uppercase turkish I (with dot) assert_equals: expected "metadata" but got "subtitles"
FAIL HTMLTrackElement.kind content attribute with lowercase turkish i (dotless) assert_equals: expected "metadata" but got "subtitles"
PASS HTMLTrackElement.kind content attribute "subtitles" 
PASS HTMLTrackElement.kind content attribute "captions" 
PASS HTMLTrackElement.kind content attribute "descriptions" 
PASS HTMLTrackElement.kind content attribute "chapters" 
PASS HTMLTrackElement.kind content attribute "metadata" 
FAIL HTMLTrackElement.kind content attribute "captions\u0000" assert_equals: expected "metadata" but got "subtitles"
PASS HTMLTrackElement.kind setting IDL attribute to "subtitles" 
PASS HTMLTrackElement.kind setting IDL attribute to "captions" 
PASS HTMLTrackElement.kind setting IDL attribute to "descriptions" 
PASS HTMLTrackElement.kind setting IDL attribute to "chapters" 
PASS HTMLTrackElement.kind setting IDL attribute to "metadata" 
PASS HTMLTrackElement.kind setting IDL attribute to "CAPTIONS" 
FAIL HTMLTrackElement.kind setting IDL attribute with uppercase turkish I (with dot) assert_equals: expected "metadata" but got "subtitles"
FAIL HTMLTrackElement.kind setting IDL attribute with lowercase turkish I (dotless) assert_equals: expected "metadata" but got "subtitles"
FAIL HTMLTrackElement.kind setting IDL attribute with \u0000 assert_equals: expected "metadata" but got "subtitles"
